Refactor: DecodeManager 역할 분리
