+++ IssueTracker/IssueTracker/Resource/API/DecodeManager.swift
@@ -8,56 +8,26 @@
 import Foundation
 
 protocol DecodeManager {
-    func requestIssues(url: String, completion: @escaping (Result<[Issue]?, NetworkError>) -> Void )
-    func getIssuesOnFailure() -> [Issue]?
+    static func decodeJson<T: Codable>(data: Data) -> T?
 }
 
 class DecodeManagerImplement: DecodeManager {
-    func requestIssues(url: String,
-                       completion: @escaping (Result<[Issue]?, NetworkError>) -> Void) {
-        guard let requestURL = URL(string: url) else {
-            return completion(.failure(.invalidUrlError))
-        }
-        var result: [Issue] = []
-        let session = URLSession.shared
 
-        session.dataTask(with: requestURL) {
-            data, response, error in
-            do {
-                guard error == nil else {
-                    return completion(.failure(.cantReachedServerError))
-                }
+    static func decodeJson<T: Codable>(data: Data) -> T? {
+        do {
+            let result = try JSONDecoder().decode(T.self, from: data)
+            return result
+        } catch {
+            guard let error = error as? DecodingError else { return nil }
 
-                if let data = data,
-                   let response = response as? HTTPURLResponse,
-                   response.statusCode == 200 {
-                    let issueResponse = try JSONDecoder().decode(IssueResponse.self, from: data)
-                    result = issueResponse.body
-                }
-            } catch {
-                completion(.failure(.invalidJsonError))
+            switch error {
+            case .dataCorrupted(let context):
+                print(context.codingPath, context.debugDescription, context.underlyingError ?? "", separator: "\n")
+                return nil
+            default :
+                return nil
             }
         }
-
-        return completion(.success(result))
     }
-    // MARK: - Local에 JSON 파일로 [Issue] 를 로딩
-    func getIssuesOnFailure() -> [Issue]? {
-        let fileName: String = "Issue"
-        let extensionType = "json"
 
-        guard let fileLocation = Bundle.main.url(
-            forResource: fileName,
-            withExtension: extensionType
-        ) else { return nil }
-
-        do {
-            let jsonData = try Data(contentsOf: fileLocation)
-            let issueList = try JSONDecoder().decode(IssueResponse.self, from: jsonData)
-            return issueList.body
-        } catch {
-            // TODO: - error 처리
-            return nil
-        }
-    }
 }
